Feature:
Separating classes into files to comply with SOLID principles.
