+++ app.rb
@@ -1,107 +1,10 @@
 require_relative 'main'
+require_relative 'book'
+require_relative 'person'
+require_relative 'rental'
+require_relative 'student'
+require_relative 'teacher'
 
-class Person
-  attr_reader :id
-  attr_accessor :name, :age, :rentals
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
-    super()
-    @id = rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-    @rentals = []
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
-  end
-
-  def correct_name
-    @name
-  end
-
-  def add_rentals(book, date)
-    @rentals << Rental.new(date, book, self)
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
-
-class Student < Person
-  attr_reader :classroom
-
-  def initialize(age, name, classroom, parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    '¯(ツ)/¯'
-  end
-
-  def classroom=(classroom)
-    @classroom = classroom
-    classroom.students.push(self) unless classroom.students include?(self)
-  end
-end
-
-class Classroom
-  attr_accessor :label
-  attr_reader :students
-
-  def initialize(label)
-    @label = label
-    @students = []
-  end
-
-  def add_student(student)
-    @students << student
-    student.classroom = self
-  end
-end
-
-class Teacher < Person
-  def initialize(age, name, specialization, parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
-
-class Book
-  attr_accessor :title, :author, :rentals
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-    @rentals = []
-  end
-
-  def add_rental(person, date)
-    @rentals << Rental.new(date, self, person)
-  end
-end
-
-class Rental
-  attr_accessor :date, :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @book = book
-    @person = person
-
-    book.rentals << self
-    person.rentals << self
-  end
-end
 
 class App
   attr_accessor :books, :people, :rentals
