+++ classroom.rb
@@ -1,0 +1,17 @@
+require_relative 'person'
+require_relative 'student'
+
+class Classroom
+  attr_accessor :label
+  attr_reader :students
+
+  def initialize(label)
+    @label = label
+    @students = []
+  end
+
+  def add_student(student)
+    @students << student
+    student.classroom = self
+  end
+end
