+++ person.rb
@@ -1,0 +1,33 @@
+require_relative 'rental'
+
+class Person
+  attr_reader :id
+  attr_accessor :name, :age, :rentals
+
+  def initialize(age, name = 'Unknown', parent_permission: true)
+    super()
+    @id = rand(1..1000)
+    @name = name
+    @age = age
+    @parent_permission = parent_permission
+    @rentals = []
+  end
+
+  def can_use_services?
+    of_age? || @parent_permission
+  end
+
+  def correct_name
+    @name
+  end
+
+  def add_rentals(book, date)
+    @rentals << Rental.new(date, book, self)
+  end
+
+  private
+
+  def of_age?
+    @age >= 18
+  end
+end
