+++ rental.rb
@@ -1,0 +1,15 @@
+require_relative 'book'
+require_relative 'person'
+
+class Rental
+  attr_accessor :date, :book, :person
+
+  def initialize(date, book, person)
+    @date = date
+    @book = book
+    @person = person
+
+    book.rentals << self
+    person.rentals << self
+  end
+end
